requirements.txt
asgiref==3.3.2
boto3==1.17.45
botocore==1.20.45
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-embed-video==1.4.0
django-heroku==0.3.1
django-storages==1.11.1
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
